Update: Cleaning Code & Plan
