+++ src/components/HeartButton.tsx
@@ -1,0 +1,51 @@
+import { useSession } from "next-auth/react";
+import { VscHeartFilled, VscHeart } from "react-icons/vsc";
+import { IconHoverEffect } from "./IconHoverEffect";
+
+type HeartButtonProps = {
+  onClick: () => void;
+  isLoading: boolean;
+  likedByMe: boolean;
+  likeCount: number;
+};
+
+export function HeartButton({
+  isLoading,
+  onClick,
+  likedByMe,
+  likeCount,
+}: HeartButtonProps) {
+  const session = useSession();
+  const HeartIcon = likedByMe ? VscHeartFilled : VscHeart;
+
+  if (session.status !== "authenticated") {
+    return (
+      <div className="mb-1 mt-1 flex items-center gap-3 self-start text-gray-500">
+        <HeartIcon />
+        <span>{likeCount}</span>
+      </div>
+    );
+  }
+  return (
+    <button
+      disabled={isLoading}
+      onClick={onClick}
+      className={`group -ml-2 flex items-center gap-1 self-start transition-colors duration-200 ${
+        likedByMe
+          ? "text-red-500"
+          : "text-gray-500 hover:text-red-500 focus-visible:text-red-500"
+      }`}
+    >
+      <IconHoverEffect red>
+        <HeartIcon
+          className={`transition-colors duration-200 ${
+            likedByMe
+              ? "fill-red-500"
+              : "fill-gray-500 group-hover:fill-red-500 group-focus-visible:fill-red-500"
+          }`}
+        />
+      </IconHoverEffect>
+      <span>{likeCount}</span>
+    </button>
+  );
+}
