+++ src/server/api/root.ts
@@ -1,6 +1,7 @@
 import { spitRouter } from "~/server/api/routers/spit";
 import { createTRPCRouter } from "~/server/api/trpc";
 import { profileRouter } from "./routers/profile";
+import { bigramRouter } from "./routers/bigram";
 
 /**
  * This is the primary router for your server.
@@ -10,6 +11,7 @@ import { profileRouter } from "./routers/profile";
 export const appRouter = createTRPCRouter({
   spit: spitRouter,
   profile: profileRouter,
+  bigram: bigramRouter,
 });
 
 // export type definition of API
